rename vaultActionsLogicFactory.ts.ts to vaultActionsLogic.ts
